backend/requirements.txt
fastapi
uvicorn
sqlmodel
apscheduler
httpx
python-multipart
python-dotenv
pydantic-settings
